ref: Use rollup API to build
