+++ package.json
@@ -62,6 +62,7 @@
         "sass": "^1.77.6",
         "ts-jest": "^29.1.5",
         "ts-node": "^10.9.2",
+        "tsx": "^4.16.2",
         "turbo": "^2.0.4",
         "typescript": "^5.5.2",
         "web-streams-polyfill": "^4.0.0"
