+++ packages/highlight/package.json
@@ -45,7 +45,7 @@
     "scripts": {
         "clean": "rm -rf lib && mkdir lib",
         "build": "npm run clean && npm run js && npm run css",
-        "js": "cp src/index.d.ts lib && cp dist/index.js lib && rollup -c ../../rollup.config.cjs",
+        "js": "cp src/index.d.ts lib && cp dist/index.js lib && tsx ../../scripts/compile",
         "css": "sass --no-source-map --style compressed src/styles/index.scss lib/styles/index.css",
         "release": "npm publish --access public"
     }
