+++ packages/search/package.json
@@ -44,7 +44,7 @@
     "scripts": {
         "clean": "rm -rf lib && mkdir lib",
         "build": "npm run clean && npm run js",
-        "js": "cp src/index.d.ts lib && cp dist/index.js lib && rollup -c ../../rollup.config.cjs",
+        "js": "cp src/index.d.ts lib && cp dist/index.js lib && tsx ../../scripts/compile",
         "test": "jest --config ../../jest.config.js",
         "release": "npm publish --access public"
     }
