+++ packages/theme/package.json
@@ -43,7 +43,7 @@
     },
     "scripts": {
         "clean": "rm -rf lib && mkdir lib",
-        "build": "npm run clean && cp src/index.d.ts lib && cp dist/index.js lib && rollup -c ../../rollup.config.cjs",
+        "build": "npm run clean && cp src/index.d.ts lib && cp dist/index.js lib && tsx ../../scripts/compile",
         "test": "jest --config ../../jest.config.js",
         "release": "npm publish --access public"
     }
